filter-setup/recognition-service/requirements.txt
numpy==1.26.4
torch==2.2.1
pillow==10.2.0
torchvision==0.17.1
opencv-python==4.9.0.80
faiss-cpu==1.7.4
imagecodecs==2024.1.1
loguru==0.7.2
httpx==0.27.0